// frontend/src/client/types.gen.ts
export type UserCreate = {
    email: string;
    password: string;
    is_active?: (boolean | null);
    is_superuser?: (boolean | null);
    is_verified?: (boolean | null);
    full_name?: (string | null);
};